Support setting resolution via argument
This works for Firefox and Chrome currently.
The static resolution previously set for chrome headless
is now gone and if no resolution is set, the default is used.
The argument used is "-r WIDTHxHEIGHT", for example "-r 1024x800"

Edge support might be added soon, but for safari I have not found
any indication that this (or headless) is supported.