Keep looking for element with waitfor-primitives when then element does not exist.
Some waitfor statements already supported this but now
all should have support.
If it is not there and we have negative search on, we
will succeed if the element is not present. Still not
set it stone if that is the best or if we should keep
on looking for the element for the full extent of time
and only bail out early if the attribute/css/etc.. is
incorrect.